+++ tests/WPMocks.php
@@ -1860,6 +1860,12 @@ namespace {
 		}
 	}
 
+	if ( ! function_exists( '__' ) ) {
+		function __( $text, $domain = 'default' ) {
+			return $text;
+		}
+	}
+
 	if ( ! function_exists( 'esc_html__' ) ) {
 		function esc_html__( $text, $domain = 'default' ) {
 			return htmlspecialchars( (string) $text, ENT_QUOTES, 'UTF-8' );